requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.0
pydantic==2.8.0

# Database
psycopg2-binary==2.9.9

# Queue
celery[redis]==5.4.0
redis==5.0.8

# HTTP & Scraping
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.3.0

# Browser automation
playwright==1.47.0
playwright-stealth==1.0.6

# LLM
anthropic==0.34.0

# Notifications
resend==2.3.0

# Utils
python-dateutil==2.9.0
structlog==24.4.0
